requirements.txt
python-telegram-bot==12.7
pymongo
dnspython
emoji